+++ CHANGELOG.md
@@ -9,6 +9,7 @@ Upcoming changes compared to [NetCoreDbg](https://github.com/Samsung/netcoredbg)
 - Added case-insensitive file name collision for all OSes.
 - Added `allowImplicitFuncEval` configuration option support (analog MSVS option: `Enable property evaluation and other implicit function calls`) https://github.com/OmniSharp/omnisharp-vscode/issues/3173
 - Added initial implementation for methods parameters output in stacktrace.
+- Added CLR native frames output in stacktrace.
 
 #### Changed
 - Replaced VSCode to DAP (variables, class names, tests, etc).